requirements.txt
ipykernel 
flask 
psycopg2-binary
transformers
torch 
flask-bcrypt
requests
streamlit
pandas